requirements.txt
anthropic>=0.40.0
boto3>=1.28.0
streamlit>=1.32.0
python-docx>=1.2.0
weasyprint>=62.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.0.0
pymupdf>=1.23.0
